separated club into clubcard.tsx
