+++ client/src/pages/clubs.tsx
@@ -14,28 +14,12 @@ import CreateClub from "../components/CreateClub";
 import { OptionsButton } from "../components/OptionsButton";
 import { DeleteEntity } from "../components/DeleteEntity";
 import { WarningIcon } from "@chakra-ui/icons";
+import { ClubCard } from "../components/ClubCard";
 
 interface Props {}
 
 const Clubs: React.FC<Props> = ({}) => {
   const [{ data, fetching }] = useClubsQuery();
-  const [, deleteClub] = useDeleteClubMutation();
-  const [{ data: userData }] = useMeQuery();
-
-  const isAuthorised = (club: Club) => {
-    return club.admins
-      .map((admin) => admin.username)
-      .includes(userData.me?.username);
-  };
-
-  const handleDelete = async (id: number): Promise<string | null> => {
-    const { error } = await deleteClub({ id });
-    if (error) {
-      console.log(error);
-      return error.message;
-    }
-    return null;
-  };
 
   if (!fetching && !data) {
     return <div>No data...</div>;
@@ -48,35 +32,9 @@ const Clubs: React.FC<Props> = ({}) => {
   return (
     <Layout>
       <CreateClub />
-      {data.clubs.map((e: Club) => {
-        console.log(e);
-        return (
-          <Card key={e.id}>
-            <Box display='flex' justifyContent='space-between'>
-              <Box>
-                <Heading>{e.name}</Heading>
-                {e.admins.map((admin) => (
-                  <Text key={admin.username}>Owner: {admin.username}</Text>
-                ))}
-                <Text>Email: {e.email}</Text>
-                <Text>{e.description}</Text>
-              </Box>
-              {/* <Text>{e.host.username}</Text> */}
-              <OptionsButton>
-                {isAuthorised(e) ? (
-                  <>
-                    <DeleteEntity
-                      handleDelete={() => handleDelete(e.id)}
-                      entityName={"Club"}
-                    />
-                  </>
-                ) : (
-                  <MenuItem icon={<WarningIcon />}>Report</MenuItem>
-                )}
-              </OptionsButton>
-            </Box>
-          </Card>
-        );
+      {data.clubs.map((club: Club) => {
+        console.log(club);
+        return <ClubCard key={club.id} club={club} />;
       })}
     </Layout>
   );
